requirements.txt
pytest==8.2.0
selenium==4.21.0
webdriver-manager==4.0.1
allure-pytest==2.13.5
